requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-pinecone
google-genai
pinecone
pandas
numpy
google-generativeai